requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.2

# Vector Store
chromadb==0.4.18

# Document Processing
langchain==0.1.0
langchain-community==0.0.10
langchain-text-splitters==0.0.1

# Utilities
numpy==1.26.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# UI (Bonus)
streamlit==1.29.0
gradio==4.8.0
